refactor: extract variable resolution functions to distri-types
Move pure functions (extract_vars, extract_vars_from_value, substitute_string,
substitute_value) from distri-core to distri-types::resolve, re-exporting them
from distri-core to preserve the existing API.
