+++ distri-types/src/lib.rs
@@ -62,3 +62,5 @@ pub use ui_tool_renderers::*;
 
 mod client_config;
 pub use client_config::DistriConfig;
+
+pub mod resolve;
